Extract Dockerfile validation diagnostics partial
